+++ CHANGELOG.md
@@ -10,12 +10,12 @@ Author: - M.Scheepers
 
 ### Features
 - added `$(!OffsetDateTime)` keyword with similar semantics as `$(!Instant)` keyword.
+- added `RequestTimeMatcher` extension that provides regular expression matching against UTC request time.
 
 ### Improvements
 - added support for custom `X-Rps-TraceId` header to callback definitions.
 - callback-simulator now persists callback definitions in the file system to reduce memory footprint of scheduled callbacks for load test scenarios to avoid OOM errors.   
 - added ability to configure callback-simulator's thread pool size using `SCHEDULED_THREAD_POOL_SIZE` environment variable so that it doesn't become a bottle neck in load test scenarios.  
-
 
 ### Fixes
 - none